extract logic to composable and remove intermediate props/vars
